Separate the implementations into their own category files
